// src/app/core/models/test.ts
import { Assessment } from "./assessment";
import { Center } from "./center";
import { DisclaimerAccept } from "./disclaimerAccept";
import { Irregularity } from "./irregularity";
import { KeyPressTracking } from "./keyPressTracking";
import { Language } from "./language";
import { RandomOtp } from "./randomOtp";
import { Screenshot } from "./screenshot";
import { Grade } from "./grade";
import { StudentTestLog } from "./studentTestLog";
import { StudentTest } from "./studentTest";
import { Subject } from "./subject";
import { TestCategory } from "./testCategory";
import { TestQuestion } from "./testQuestion";
import { TestSecurityLevel } from "./testSecurityLevel";
import { TestType } from "./testType";
import { UploadedSourceDocument } from "./uploadedSourcDocument";
import { UserDocumentAnswer } from "./userDocumentAnswer";
import { UserDocumentAnswersBackup } from "./userDocumentAnswersBackup";
import { UserScannedImage } from "./userScannedImage";

export interface Test {
    name: string;
    surname: string;
    idNumber: string;
    studentNo: string;
    emailAddress: string;
    contactNo: string;
    gradeId: number;
    examNo: string;
    regionId: string;
    id: number;
    paperExpiryDate: Date | null;
    code: string | null;
    testName: string | null;
    testIntro: string | null;
    testDuration: string | null;
    testCreated: string | null;
    examDate: Date| null;
    languageId: number | null;
    dateModified: string | null;
    modifiedBy: string | null;
    testDocBase64:string | null;
    tts: boolean | null;
    workOffline: boolean | null;
    testSecurityLevelId: number | null;
    answerScanningAvailable: boolean | null;
    subjectId: number | null;
    sectorId: number | null;
    examId: number | null;
    testTypeId: number | null;
    testCategoryId: number | null;
    alternateTestId: number | null;
    centerId: number | null;
    oldSubjectId: number | null;
    oldSectorId: number | null;
    oldCenterId: number | null;
    assessments: Assessment[];
    center: Center | null;
    disclaimerAccepts: DisclaimerAccept[];
    irregularities: Irregularity[];
    keyPressTrackings: KeyPressTracking[];
    language: Language | null;
    randomOtps: RandomOtp[];
    screenshots: Screenshot[];
    sector: Grade | null;
    studentTestLogs: StudentTestLog[];
    studentTests: StudentTest[];
    subject: Subject | null;
    testCategory: TestCategory | null;
    testQuestions: TestQuestion[];
    testSecurityLevel: TestSecurityLevel | null;
    testType: TestType | null;
    uploadedSourceDocuments: UploadedSourceDocument[];
    userDocumentAnswers: UserDocumentAnswer[];
    userDocumentAnswersBackups: UserDocumentAnswersBackup[];
    userScannedImages: UserScannedImage[];
}